last fix

+++ Server.java
@@ -58,7 +58,7 @@ public class Server
     		double assAngle = assassin.angleFinder(vicLat, vicLng);
     		assassin.give("angle " + assAngle);
 			
-			if (distFrom(vicLat, vicLng, assLat, assLng) < 2)
+			if (distFrom((float)vicLat, (float)vicLng, (float)assLat, (float)assLng) < 2)
 				victim.give("hurt");
 			
 			victim.give("health");
